backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
scikit-learn
prophet
python-multipart

uvicorn[standard]
tensorflow
Pillow
